+++ cmd/gen/makefile/flag.go
@@ -16,21 +16,29 @@ const (
 )
 
 type flag struct {
-	Flavour  string
-	Language string
+	Flavours gen.FlavourSlice
+	Language gen.Language
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
-	cmd.Flags().StringVarP(&f.Flavour, flagFlavour, "f", "", fmt.Sprintf(`The type of project that you want to generate the Makefile for. Possible values: <%s>`, strings.Join(gen.AllFlavours(), "|")))
-	cmd.Flags().StringVarP(&f.Language, flagLanguage, "l", "", fmt.Sprintf(`The programming language of project that you want to generate the Makefile for. Possible values: <%s>`, strings.Join(gen.AllLanguages(), "|")))
+	cmd.Flags().VarP(gen.NewFlavourSliceFlagValue(&f.Flavours, []gen.Flavour{}), flagFlavour, "f", fmt.Sprintf(`List of types of project that you want to generate the Makefile for. Possible values: <%s>`, strings.Join(gen.AllFlavours(), "|")))
+	cmd.Flags().VarP(gen.NewLanguageFlagValue(&f.Language, gen.Language("")), flagLanguage, "l", fmt.Sprintf(`The programming language of project that you want to generate the Makefile for. Possible values: <%s>`, strings.Join(gen.AllLanguages(), "|")))
 }
 
 func (f *flag) Validate() error {
-	if !gen.IsValidFlavour(f.Flavour) {
-		return microerror.Maskf(invalidFlagError, "--%s must be one of <%s>", flagFlavour, strings.Join(gen.AllFlavours(), "|"))
+	if len(f.Flavours) == 0 {
+		return microerror.Maskf(invalidFlagError, "--%s must not be empty", flagFlavour)
 	}
-	if !gen.IsValidLanguage(f.Language) {
-		return microerror.Maskf(invalidFlagError, "--%s must be one of <%s>", flagLanguage, strings.Join(gen.AllLanguages(), "|"))
+	if len(f.Language) == 0 {
+		return microerror.Maskf(invalidFlagError, "--%s must not be empty", flagLanguage)
+	}
+
+	if f.Flavours.Contains(gen.FlavourCLI) && f.Language != gen.LanguageGo {
+		return microerror.Maskf(
+			invalidFlagError,
+			"flavour %q is supported only for language %q",
+			gen.FlavourCLI, gen.LanguageGo,
+		)
 	}
 
 	return nil