+++ cmd/gen/workflows/flag.go
@@ -17,17 +17,17 @@ const (
 
 type flag struct {
 	CheckSecrets bool
-	Flavour      string
+	Flavours     gen.FlavourSlice
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
-	cmd.Flags().StringVarP(&f.Flavour, flagFlavour, "f", "", fmt.Sprintf(`The type of project that you want to generate the workflows for. Possible values: <%s>`, strings.Join(gen.AllFlavours(), "|")))
+	cmd.Flags().VarP(gen.NewFlavourSliceFlagValue(&f.Flavours, gen.FlavourSlice{}), flagFlavour, "f", fmt.Sprintf(`The type of project that you want to generate the workflows for. Possible values: <%s>`, strings.Join(gen.AllFlavours(), "|")))
 	cmd.Flags().BoolVar(&f.CheckSecrets, flagCheckSecrets, true, "If true, also generate a secret-scanning workflow. Possible values: true (default), false.")
 }
 
 func (f *flag) Validate() error {
-	if !gen.IsValidFlavour(f.Flavour) {
-		return microerror.Maskf(invalidFlagError, "--%s must be one of <%s>", flagFlavour, strings.Join(gen.AllFlavours(), "|"))
+	if len(f.Flavours) == 0 {
+		return microerror.Maskf(invalidFlagError, "--%s must not be empty", flagFlavour)
 	}
 
 	return nil