+++ pkg/gen/input/makefile/internal/file/makefile_gen_go_mk.go
@@ -10,7 +10,6 @@ func NewMakefileGenGoMkInput(p params.Params) input.Input {
 		Path:         "Makefile.gen.go.mk",
 		TemplateBody: makefileGenGoMkTemplate,
 		TemplateData: map[string]interface{}{
-			"IsFlavourApp": params.IsFlavourApp(p),
 			"IsFlavourCLI": params.IsFlavourCLI(p),
 			"Header":       params.Header("#"),
 		},